requirements1.txt
aiofiles==0.8.0
anyio==3.6.1
APScheduler==3.9.1
arrow==1.2.2
asgiref==3.5.2
asyncio-dgram==2.1.2
bbcode==1.1.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
click-plugins==1.1.1
colorama==0.4.5
cookiecutter==1.7.3
cycler==0.11.0
dnspython==2.2.1
emoji==2.0.0
fastapi==0.78.0
feedparser==6.0.10
fonttools==4.37.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
imageio==2.21.1
Jinja2==3.1.2
jinja2-time==0.2.0
kiwisolver==1.4.4
loguru==0.6.0
MarkupSafe==2.1.1
matplotlib==3.5.3
mcstatus==9.2.0
msgpack==1.0.4
multidict==6.0.2
nb-cli==0.6.7
nonebot-adapter-mirai2==0.0.14
nonebot-adapter-onebot==2.1.1
nonebot-plugin-apscheduler==0.1.3
nonebot-plugin-code==0.0.4
nonebot-plugin-date-name==1.0
nonebot-plugin-directlinker==3.0.0
nonebot-plugin-emojimix==0.1.8
nonebot-plugin-eventdone==0.1.8
nonebot-plugin-imageutils==0.1.8.1
nonebot-plugin-kawaii-robot==3.2.0
nonebot-plugin-mcstatus==0.1.6
nonebot-plugin-petpet==0.3.10
nonebot-plugin-repeater==3.1.4
nonebot2==2.0.0b4
numpy==1.23.2
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pilmoji==2.0.2
poyo==0.5.0
prompt-toolkit==3.0.30
pydantic==1.9.1
pyfiglet==0.8.post1
pygtrie==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
sgmllib3k==1.0.0
shodan==1.28.0
six==1.16.0
sniffio==1.2.0
some-package==0.1
soupsieve==2.3.2.post1
starlette==0.19.1
text-unidecode==1.3
tomlkit==0.10.2
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
XlsxWriter==3.0.3
yarl==1.7.2
